renamed validator package. removed unnecessary import.
