+++ jparest/src/test/resources/test-context.xml
@@ -30,5 +30,5 @@
     <bean id="persistenceExceptionTranslationPostProcessor" class="org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor" />
 
     <bean class="uk.gov.homeoffice.digital.sas.jparest.config.ObjectMapperConfig"/>
-    <bean class="uk.gov.homeoffice.digital.sas.jparest.utils.EntityValidator"/>
+    <bean class="uk.gov.homeoffice.digital.sas.jparest.validation.EntityValidator"/>
 </beans>
